Free xref bytes instead of object on allocation failure.
If we fail to allocate space for the entries in an xref, don't try to
countdown the object because we haven't yet constructed enough of the
xref table to qualify as an object. Just free the bytes.

This requires a separate fix in the chunk memory allocator, without
that the allocation size gets truncated from a size_t to a uint (32 bits)
and the returned pointer is insufficiently large for the following
memset, leading to a eg fault.

File 0698279_13995_poc.pdf from the bugzilla PDF file collection.