Bug 703404: XPS to pdfwrite patterns
Patterns that are filled with glyphs and then subsequently
filled with another pattern (Visual brush in XPS) was not
working when going to pdfwrite due to the fact that
gs_text_begin gets called by the XPS interpreter
before we have finished the pattern fill of the glyph.
This ends up referencing the pattern that is not done
which ends badly for high level patterns.  Instead
if we are already in a high level pattern, set the color
to just a gray.  Then after the gs_text_begin, the XPS
interpreter sets the brush to the visual brush which is
subsequently collected.   Fixing this, has revealed another
XPS to PDF bug where we are not doing shading fills of glyphs
correctly (at least when dealing with patterns).