Turn numbers greater than 64-bit into reals rather than integers
If an integer value exceeds a signed 64-bit representation we need to
turn it into a real number instead.

This was causing us to fall into a seg fault in the PNG Predictor
Flate decoding which GS avoids because it throws out the Columns
value because it is not an integer.

NB I think the PNG Predictor could do with fixing as well, it shouldn't
seg fault with ridiculous values for /Columns and I think it might still
be possible to provoke this, but I am not certain.

But this code was incorrect before anyway.

File is from the Bugzilla PDF collection:
SIGABRT-070214-235544-6.pdf