Have pdfi_string_bbox() return "advance width" (and some fixes)
Since the bounding box will not account for non-marking glyphs at the start or
end of a string, also return the accumulated advance width for the glyphs in
the string.

In addition, the function was not fully accounting for variations in the text
rendering matrix or ctm, so this should fix those problems, too.