Improve the detection of extremely large numbers as reals
This caters for negative numbers which the previous commit omitted
(oops).

This commit fixes the integer to the Ghostscript-defined maximum integer
this is not the same as the maximum integer on the platform!

For example, on a 64-bit Windows build the maximum integer is still given
as 2^32-1, not 2^64-1. This matches the behaviour of the Ghostscript
interpreter, and works around a number of problems with the PNG
Predictor and FlateDecode.