Fix bug 703245: bitmap size exceeds buffer.
The pdf14 compositor changes the color depth to 8 for a SMask, but the
compositor actions were not being put in all the bands needed. In this
case the copy_color_alpha was output from the interpolation logic because
the image gridfitting expanded the image. Fix it by expanding the top and
bottom limits for the "temp_cropping" which controls the RECT loop writing
the compositor info to the bands.