Initialise some PDF-specific portions of the graphics state
The text rendering matrix and text line matrix, which are PDF-specific,
were not being initialised when the graphics state was created. This
should not be a problem in normal operation because the PDF interpreter
should always initialise them before use.

However, during development on pdfi it did show up so this commit
initialises the matrices to the identity.