pdfi_string_bbox: cope with negative coordinates
Originally, pdfi_string_bbox() set the current point to the origin before
drawing the string to measure the results. But since devices (including the
bbox device) clip to their width/height, any glyph features that extended
"back" (unlikely!) and/or "down" were clipped and thus not included in the
bounding box.

That simplified things, but obviously didn't work adequately.

So this uses a (let's call it!) heuristic to move the starting point to an
offset which should avoid such problems. Then undoes that for the advance width
value.