Fix buffer overrun in ASCIIHexEncode
This exhibited as a memory corruption error (buffer overrun) in pdfi.

The ASCIIHexEncode filter is required to write linfeed characters to the
output at least every 80 (output characters). Ghostscript chooses to do
so every 64 characters. So every 64 characters the filter emits a
linefeed, crucially without checking whether there is room in the
output buffer.

In order to determine how many linefeed characters we are going to need
to write into any given output buffer we divide the number of characters
in the output buffer by 65 (64 characters + linefeed), taking into
account any character slaready emiitted (stored as 'count' in the stream
state). We subtract that from the size of the output buffer, in order
to ensure that there is enough spare space for the linefeeds.

Unfortunately, this is not sufficient. Consider an output buffer 194
bytes in size, on the first pass. We divide that by 65, which tells us
that 2 complete buffers will fit, so we need to reserve space in the
output for 2 bytes. We subtract 2 from the count, giving 192, We then
proceed to write the output data. We write 64 bytes of data, then a
linefeed, then 64 bytes of data, then a linefeed, then 64 bytes of data.
Total 194 bytes.

At this point we have exhausted the output buffer, but because we have
written 64 bytes, we unconditionally write a linefeed. This is written
one byte beyond the end of the buffer.

There are many ways to address this but I've chosen to change the
divisor from 65 to 64. This may mean that occasionally we will write one
input character (2 output characters) less than the maximum possible
to the output, but it will prevent the possibility of a buffer overrun.