Improve handling of text rend4ring modes
When executing text rendering modes which in effect use 'charpath' in
order to stroke/fill/clip text we must not alter any existing path in
the graphics state. So we need to gsvae/grestore around the operation
so that the existing path is unaffected.

However, the current point *should* be updated, so in effect we need to
append a 'moveto' to any extant path in the graphics state after we
execute any of these operations. We weren't doing that while left the
current point where it was before the text operation.

This commit gets the current point and executes a moveto after the text
operation, which improves the situation.

However I do not believe this is sufficient, I suspect that changes in
the text matrix can leave the current point incorrect. In addition we
still don't apply any of the clip rendering modes. Finally, the
pdfi_show() function is too large and unwieldy, it needs splitting into
smaller functions, possibly one per rendering mode.

So, still work to do here.