Bug 703265: Tweak create_compositor device method.
Update create_compositor device method, so that it always returns
the compositor device (or the leaf device if there is no specific
compositor device). The device now returns 1 if we created a
compositor device to wrap the given device.

This should enable us to identify exactly the cases where forwarding
devices need to update which device they forward to.

In particular, this allows us to remove the horribly fragile code
in apply_create_compositor in gxclrast.c, and to ensure that we
correctly identify the 'new compositor' case.

This avoids us sending stuff to the wrong device, and having to
cope with a slew of warnings.