Change pdfi_obj_dict_to_stream() to optionally convert object number
We have two different use cases.  In one of them we are in the middle of
creating the object and need to convert a dict to stream.  In this case
we want the stream to have the dict's object number.

In the other case we are just creating a temporary stream and sticking
the existing dict in it, in which case we don't want to muck with the
object numbers.

This fixes a warning about freeing an object that is in the cache,
because mucking with object numbers when it's in the cache is a bad idea.