Rewrite text handling for render modes other than 0
The old code used the same charpath for stroke and fill, which it should
not have, didn't update the current point for modes other than 0, didn't
perform clip properly, and especially didn't handle modes involving
clipping when preserving the text rendering modes (because pdfwrite is
historically crippled in this regard).

This code fixes a lot of files, improves a number of others, but does
cause one file to exit with errors. It always should have, because it
tries to set the current point approximately 2,600 miles away, but we
previously swallowed the error, now we don't.