Bug 703273: Prioritise Nonsymbolic over Symbolic font flag
The font in this PDF has it descriptor flags set to 36, that is bit 3 (Symbolic)
and bit 6 (Nonsymbolic) are both set. Since the two flags are mutually exclusive
this is clearly broken!

Previously, we only test the Symbolic flag, since !Symbolic should imply
Nonsymbolic, which caused us to use the wrong encoding scheme.

Now, we'll act on the Symbolic flag iff the Nonsymbolic flag is not set.

(Originally a mupdf report, that also exhibited in Ghostscript).