Fix compositor operation with subclassing.
Compositors/PS2Write/Subclassing/Clist interact badly, due to the need
for the subclass code to 'tightly' wrap the original device, rather
than wrapping the created compositor.

Use a new gxdso to handle the insertion of the device lower down the
device chain.

This frees the subclass device of the need to know how to insert itself
into different types of device.

Unfortuntely, the pdf14clist devices, when popped, assume that they
tightly wrap the actual clist device. This causes SEGVs when there
is actually a subclass device in the way.

Accordingly, we update that code to find the correct clist device
in the stack, by using a new gxdso added for this purpose.