Make test for DEBUG of freeing object that is in cache more selective
Also test to make sure the object pointer is the same as the one in
the cache.  This detects cases where we never cached the object in
question, but it was also in the cache.

For example this happens in pdfi_dereference() on broken file
tests_private/pdf/sumatra/424_-_page_6_does_not_render.pdf
when it was dealing with a free'd object entry (the pdfi_pop() at L785).

Also fix debugging to be compatible with REFCNT_DEBUG
When REFCNT_DEBUG enabled, it was declaring ctx1 in the middle of a block,
so I moved it up to where it's valid.