Tweak replace_cache_entry() to avoid triggering warning message
This code could trigger a spurious warning about freeing an object
that was in the cache.  I re-ordered things to avoid the situation.

Basically instead of decrementing the object directly from the cache,
save a reference to it and do the decrement after it's no longer in
the cache.

This seems hacky, but it's kind of a special case, since it's code
that's manipulating the cache itself.

Sample: tests_private/comparefiles/Bug690113.pdf