Fix erasepage_optimization device to allow for other subclass devices.
The epo device assumed that it was the current device, then checked the child
of that device to decide if it could optimize. If another subclass device was
the current device the decision could be invalid.

Note that the comparison of the child's fillpage proc is not ideal, but the
change to that is more involved and risky, and left for a subsequent change.