Fix a reference count with TrueType /Encodings
This looks like a simple oversight/typo. If we fail to get an /Encoding
from a font dictionary we create a name object /MacRomanEncoding in
order to build an internal Encoding object below.

Because pdfi_name_alloc() now returns an object with a reference count
of 0 we need to count it up. But missing parentheses for the if meant
that we counted up the Encoding object even if we had pulled one out
of the font dictionary, and that already has a reference count of 1.

Counting it up unconditionally led to a memory leak.