Fix uninitalised variable
It looks like 'pdfmark' is initialised, because it is se to the return
value from pdfi_device_check_param_bool(), but that function uses
param_read_bool() and returns the parameter it reads.

The problem is that if the device doesn't understand/support the
requested parameter it does not set the return value, because it can't
know what a sensible default is. So we return an uninitialised value.

Picked up by Chris while running a valgrind check.