Do not abort image if the /Mask is invalid
From a customer file with an image where the Mask is given as

/Mask ([2 2])

Clearly this should not be a string, Acrobat apparently ignores the
/Mask and renders the entire image, we aborted the image and carried on
with the remainder of the content.

This commit alters the behaviour so that we draw the image and note the
error. Added code to emit a warning on exit.