Bug 703324 - restore fill state after fill+stroke path operation
The file in bug 703324 fills and strokes a rectangle ('B' operator)
and sets the fill constant alpha to 1, and the stroke constant alpha
to 0, so that the stroke is not drawn. Pointless but that's what it
does.

It then draws some text.

Now the clist maintains an 'opacity' member, which is set and reset by
(amongst other things) looking at the fill and stroke alpha values. When
we do a fill+stroke we set the opacity to the fill alpha, and do the
fill, then we set it to the stroke alpha and draw the stroke.

We then do an 'image_fill_masked' operation to draw the text bitmaps.
This device method does not take a graphics state as a parameter and so
is unable to set the opacity in the device based on the fill alpha. It
simply uses whatever opacity is current at the time. Because the
fill+stroke left this at the 'stroke' alpha, this is applied to the text
which (because the alpha is 0) is then not drawn.

In addition the graphics state fill constant alpha value was left with
the wrong value, but fixing that wasn't sufficient. However we should
do both, and this commit resets the fill constant alpha in the graphics
state and recalculates the 'opacity'.