Fix ordering of operators in FitPage
Bug #703685 "Crash when using -dPDFFitPage"

The 'cvi' in pdf_PDF2PS_matrix in pdf_main.ps here was meant (I think)
to apply to the /Rotate value from the page dictionary, but instead
was applied to the divisor, 90, which isn't needed because 90 is
obviously already an integer.

If the PDF file (illegally) had a floating point number then the cvi
should convert it to an integer, so we need to move the cvi so that it
applies to the /Rotate value.