Bug 703687: Fix logic for alpha bits in do_fill_stroke()
We can only do the GraphicsAlphaBits anti-aliasing if we are drawing with a
"pure" color (or a DeviceN color). Basically, not a pattern.

For fill_stroke that means checking for the fill and stroke colors, but the
conditional was erroneously using "or" rather than "and" - so if either
color was pure, we'd try to use the alpha buffer, causing incorrect results.