Fix SEGV with -sNupControl=1x2 to pbmraw -r300 Bug693331.pdf
This was due to the pdf14 getting confused during clist playback when the
'master' page contained more than one nested page with transparency. Each of
the nested pages would PUSH, then POP the pdf14 compositor, but since the pdf14
compositor remains as the current device, the first POP 'disables' the pdf14
compositor and the subsequent PUSH 'recreates' the pdf14 device in active
compositing mode. The problem was that the 'recreate' did not perform all of
the setup of color_info, cmap_procs, and device procs needed for all of the
subsequent clist playback through the compositor.