Fix stream keyword parsing
The stream parsing code was over enthusiastic about removing line feed
and carriage return characters at the end of a 'stream' keyword.

The problem is that we need a delimiter for the end of the keyword, but
a compressed/encrypted/binary stream can contain arbitrary data and can
easily start with CR/LF characters. According to the spec the stream
keyword should be followed by either a linefeed or a carriage return
and then a linefeed. Naturally we have examples which do not follow this
rule.

This code seems to work better (particularly not stripping CR/LF at the
start of a stream) and shows several progressions.