Implement the basic code for AcroForm Sig fields
If there's a Normal Appearance stream, then draw it, if there is no
Appearance, do a debug message and continue.

Synthesising an appearance for a Sig fields seems like its probably not
possible to me, but we'll leave this as a TODO for now.

Initialise a variable to silence an (incorrect) compiler warning.