Apply the CropBox as a clip
Ghostscript applies the CropBox as a clip rectangle, even when it is
not using the CropBox as the Media size (UseCropBox).

This commit implements the same behaviour for pdfi.

Add gspath2.h to ghostpdf.h to get the prototype for gs_rectclip()

Also changed a copy/paste error which was retrieving MediaBox when
UseTrimBox was specified.