Drop DynaLab special bbox stretching of metrics
For historical reasons relating to glyph caching the zchar42_set_cache
code used the minimum and maximum of the glyph bounding box and font
bounding box when setting the metrics of a glyph.

This was done for DynaLab (one of the FreeType 'tricky' font types),
because these fonts abuse the TrueType 'hinting' mechanism.
Unfortunately this results in the wrong origin 1 being sent to
setcachedevice2, and hence CDevProc when the font is vertically
oriented.

We no longer use this function for clipping, and the remaining client
pdfwrite doesn't use it for clipping but very much does use it for
gathering font metrics, so it is vital that these be correct.

Since it's only used by pdfwrite, and testing suggests that DynaLab
fonts are not a problem for this, the code has been removed here. If
it should ever prove to be required then we should check the Font to
see if it is a DynaLab font. MuPDF has a function is_dynalab() for this
and we should use that if we prove to need it.