Erasepage optimisation - fix invalid device ordering assumption
The erasepage optimisation device, when checking to see whether the
underlying device permitted optimisaton, was assuming that its own child
would always be the 'final' device. If we have a chain of subclass
devices this might not be true.

Run down the chain to the fist non-subclass device and check that
instead. In future this should be replaced by a spec_op or something
but this resolves the problem for now.