Fix crash in display device (seen with apitest.c).
When calculating the size for the display, we setup a
fake gx_device_memory to call the calculation routine
with. Because we hadn't initialised the graphics_type_tag
sometimes it would incorrectly assume there was 1 more
plane than there should be.

In calculating the required size, it would therefore add
in bitmap_raster(plane[i].depth * width) more bytes for
that plane. If plane[i].depth was negative, this would
cause the amount of memory we allocate to shrink, and
accordingly, we'd allocate too little memory.

The fix is to ensure that the fake device is always
zero'd to start with.