Fix a performance regression introduced by commit 813b5f48e8
The fix is to keep the PageUsesOverprint parameter in the device
dictionary, not in the outer dictionary, and use it from there.

It isn't understood why storing in the outer dictionary causes such a
substantial performance hit, but profiling shows that the time comes
from "dstack_find_name_by_index". The performance hit showed up with
j9_acrobat.pdf of the performance test suite (about 8%).