Automate tesseract build compatibility with clang++
On MacOS (at least), clang++ requires the addition of the -std=c++17 to enable
support of the features that tesseract uses.

Also, on Linux (at least) clang++ requires -stdlib=libstdc++ in order to
pickup a usable stdc++ library.

So, tweak configure to test whether the compiler accepts the options to do both
the above, and if so, add them to the CXXFLAGS used to build tesseract.

In addition, in configure, pickup CXXFLAGS so custom options can be set on the
configure command line.