Fix JPXDecode image issues
If the JPX header was valid, override BitsPerComponent from the image header.

sample: tests_private/comparefiles/Bug695387.pdf

Also add no_data flag and is_valid cases for JPXDecode

In the case where we are making an smask for SMaskInData of a
JPXDecode image, the smask will end up being flagged as a JPX image
when it really isn't because it has no separate header.  This catches
that case, which is exposed by the other part of this fix.

Also there are still cases where it's a JPXDecode image, but doesn't
have a valid header.  In that case we have is_valid=false, and we
don't try to use the jpx_info->bpc from it.

Sample: tests_private/comparefiles/Bug694873.pdf