Support for JPXDecode colorspaces
Partial implementation to support the JPXDecode enum colorspaces

This puts in the mapping that is in jp2_csp_dict in the gs code.
It tries to load this colorspace by name (which will fail) and
then loads the backup colorspace, which should do the same as it did
before.

Need to implement support for these in pdf_colour.c, however this
is best done.  It looks like pdfi_create_icc() supports the
GSICC_STANDARD_PROFILES, so the bottom piece is in place, just lacking
some plumbing in between.

gs code is scattered around in gs_res.ps and gs_lvl2.ps.