For pdfi-inside-Ghostscript, fix the page Group size
We use the ctx->page.Size values for the initial page Group when a
page has a Group. We want to use the actual media size of the device,
so that we cover the actual output.

However, the values passed to the compositor should be in user space,
so we need to apply the inverse transform of the CTM in order to get
the correct values.

This appears now to work correctly for rotated and offset media.