pdfwrite - fix FastWebView with multiple put_params() calls
When testing this I must have used a setup which either set the
OutputFile before setting FastWebView, or set them simultaneously. If we
have the two assignments in different calls, and the first call is not
to set OutputFile, then the assignment of FastWebView will not work.

This is because we didn't check whether OutputFile was set before
checking if it was seekable, obviously a NULL file is not seekable! This
caused us to reset 'Linearize'. The subsequent call to set OutputFile
could not re-enable Linearize, and so linearization would not be
enabled.

The fix here is simply to check that pdev->flle is not NULL before
checking if it is seekable.