Tweak use of return_error to be more useful.
In these two instances, the existing code is calling
return_error(gs_error_rangecheck), even when the functions
are being called to check what the required size is (i.e.
when gs_error_rangecheck is not actually an error).

Tweak the code so that we no longer hit breakpoints in
gs_note_error for such cases.