Add the JPX non-device colour spaces
Follows on from Nancy's commit de5b3cdcd4b2af9468fc15c9e8c20d13d163f552
this adds the named colour spaces esRGBICC, rommRGBICC, sRGBICC and
sGrayICC to the colour space code.

These are all created by a single function which picks up the standard
embedded ICC profile and creates the space from that.

The JPX Lab space, which requires a D65 white point, is manufactured
in the image handling code, since we already have colour space code to
deal with Lab spaces.

Includes a minor fix in the colour space code to check that the stream
provided to read an ICC profile is not NULL and return an error if it
is.