Don't call initgraphics before drawing annotations
When running pdfi natively something like this may eventually be
required (see the comments just above. But at the moment it makes no
difference in that case.

However when running under Ghostscript we must not call this, as we
modify the graphics state before starting pdfi, and if we call
intigraphics it will reset to the state before we did that which will
result in the annotations potentially being drawn incorrectly.

If we need to set a specific graphics state we should look at
currentgstate/setgstate; get the current graphics state when we start
up and store it in the context, then set the graphics state from that
if we need to reset.