Simplify some pdf14 device prototype code.
The existing code passes in a pointer to a temporary device structure
and a pointer to a device structure pointer that will be filled in
(almost always) to point to the temporary structure.

Simplify the code to just take a device structure pointer that is
always filled in.

Simpler is better.