Another DecodeParams 'fix'
If the number of entries in the DecodeParms array is not the same as
the number of entries in the Filter array, then ignore the DecodeParms
array.

This seems wrong, but doing so matches Acrobat. Of course if any of the
Filters uses non-default DecodeParms, they will not work correctly.