Fix Text rendering mode 2 for transparency
The fill and then stroke text rendering modes were just using the
fill and stroke rendering mode code, but that doesn't work properly if
there is any transparency involved (eg different fill and stroke
constant alpha).

However we now have a fillstroke primitive, which handles this, so we
can use that instead and get the correct results.