Bug 703752: More broken image handling
The image rendering throws an error, in this case, leaving the operand stack
polluted. Even though we later catch the error, and try to carry on, the
polluted op stack prevents the remainder of the content stream from executing.

So, wrap in the "inner" level of image rendering in a stopped context, and
clean up the stack when we're done.