Fix handling of Processed flag
Need to clear it if another gs command with an SMask happens.
This is to follow the hacky logic in the gs code.

I am not very happy with it, but the results seem to be progressions,
so let's try it for now.

Sample: tests_private/pdf/uploads/halloween.pdf