Change names from Graphics and Path to Vector, for consistency
We decided to follow the ObjectFilter subclass device's naming convention
for objects that are not Image or Text. This patch makes them all the same.

NOTE: The GS9_Color_Management.pdf document also needs to change since we
are changing the external names as well. Before going 'live' with this non-
backward compatible change (admittedly to something that has probably not
been widely used) we may consider issuing a warning for the previous option
names, or even a warning with automatic use of the new name.