Move memory devices away from static init of proc tables.
This is slightly more involved than the other devices, as
internally some memory devices rely on being able to get
the prototype for other ones, and to reference their
functions.

Accordingly, we change to a much more restricted list of
procs (just for the ones that actually differ between
memory devices) and have an internal mechanism for getting those.

attempt to simplify init of memory devices.