Deal with NULL CS in Transparency Group without causing error
Sample: tests_private/pdf/sumatra/2258_-_invalid_transparency_group_colorspace.pdf

This sample has an invalid (deleted) /CS references in the Transparency Group.
We now return a NULL object for this case, but instead of treating this as
an error, just treat it the same as if the /CS entry wasn't there at all.

NOTE: I am calling pdfi_dict_knownget() -- shouldn't this detect the
NULL and return a 0 instead of a >1?  I mean, if it got a NULL object,
that means there was an error with the object, so is it really
"known"?  I am wondering if we will find all kinds of similar
error-handling glitches caused by the current behavior?