pdfwrite - alter the way Pattern resources handle graphics state
Bug #703738 "Transparent Object Lost during pdfwrite"

The problem with this file is that the Pattern colour space does not contain
operators to set the transparency, but instead 'inherits' the transparency
at the time the Pattern PaintProc is run.

In effect the file has:

/GS2 gs			% set CA and ca to 0.11
/CS0 cs /P6 scn	% Set the current colour space to a Pattern
/GS0 gs			% set CA and ca to 1
..construct path..
f				% fill the path.

Clearly Acrobat does not apply the transparency at the time the Pattern is
set as the current colour, but instead does not evaluate the PaintProc until
the Pattern is actually used to paint the Pattern cell.

Currently pdfwrite captures the unsaved graphcis state changes at the start of
the Pattern, and writes them into the Pattern, thus causing the Pattern itsef
to have transparency, which the original does not.

However certain Quality Logic test files (in PostScript) do the equivalent:

0 setlinecap
...
setpattern
...
1 setlinecap
fill

And in these cases the pattern is drawn using the butt linecap, not round.

So the two are incompatible; for PDF we need to defer the graphics state
until the fill, and for PostScript we need to capture it as part of the
Pattern creation.

The PLRM does say :

"Because of the effects of caching, the PaintProc procedure is called
at unpredictable times and in unpredictable environments. It should depend
only on information in the pattern dictionary and should produce the same
effect every time it is called"

Basically, don't rely on the graphics state. Since the PLRM says this
I'm concluding that we can treat the behaviour for the QL files as
undefined, and simply operate as per the requirements for PDF.

This commit write out any changes to the current graphics state at the
start of any Pattern PaintProc. It then saves the current state, in a
stack of initial states (because Patterns can use Patterns), and sets
the current graphics state and pdfwrite's 'saved state' to the initial
graphics state. It then runs the PaintProc. When the PaintProc is
complete we reset the graphics state and saved state from the stack.

This has the effect of capturing any *changes* to the graphics state
in the Pattern PaintProc, and only changes actually made during the
course of the Paintproc. The graphics state at the time the Pattern
is set is correctly written out, as are any further changes before the
Pattern is used.