Apply text spacing when we previously did not
For fonts we have apply the text spacing value in with the Widths, because the
graphics library doesn't allow "replace" and "add to all" metrics in the same
operation.

But for CIDFonts, and any fonts without Widths arrays, we can instruct the
graphics library to add our extra width to the glyph for us.