Implementing transparency with text
This implements the basic Tj case, and fixes several files (plus a lot
of diffs that are too small to tell with naked eye).

Not complete yet.

I was specifically working with Tr 0 but it may work for a bunch of other
cases also.

Unfortunately the gs implementation is particularly annoying to try to
understand...