In BT only do an initial moveto if there is no currentpoint
In ET reset to a new path if there was initially no current point.

Fixes /tests_private/comparefiles/Bug692867.pdf.pdf the file has an
invalid 'eofill' because it does a lineto when there is no path. Our
creation of an initial path caused that line to become valid and be
drawn.

We now store the initial state of the current point on a BT, if it is
invalid we do a moveto so that there is a current point for the text.
When we get a ET we check the initial current point and if it was
invalid we do a newpath.