Harden the use of the ICC cs cache for Cal spaces
Originally, the cache key used was the "saveid" entry in the ref "value" union.
Since the ref object in question was a dictionary, this meant an unsigned long
interpretation of the pointer value from "dictval" in the same union.

This was unsafe for a couple of reasons: Postscript allows dictionary objects
to be reused, so it was feasible for the same dictionary object to be used to
define two different color spaces, this meaning the cache would retrieve the
wrong ICC color space.

Secondly, garbage collection results in objects (like dictval instances) moving
address. So, the same dictionary need not be at the same address, but a
different dictionary could be.

Since the other CIE based spaces use an MD5 hash of the contents of the color
spaces to create the cache key, this implements the same things, bringing the
Cal color spaces in-line with their CIE bretherin.

In addition, where we retrieve color spaces from the ICC cs cache, add in a
simple validation of the number components before going ahead and using the
cached color space.

This came out of investigating a *long* standing indeterimate cluster test file
tests_private/comparefiles/p2b-100.pdf and we fervently hope that this will fix
that indeterminacy.