Need to do q/Q for shading operator to keep transparency state correct
The q/Q operators will do the gs_pop_transparency_state() if
page has transparency.  If that's not done, then things get out of sync.

Basically, in this file the shading command set a mask.  Even though
the shading was done in its own q/Q in the source file, it still
wasn't correct.

There is a mask stack maintained in the transparency code and
pgs->trans->xstate_change needs to be tested (in
gs_pop_transparency_state()) when it's still true in order to keep the
transparency stack correct.  Doing a pdfi_grestore() or gs_grestore()
instead will miss the opportunity.

The bug was that the transparency machinery thought there was still a mask
set later when it did the form, even though the mask had nothing to do with
the current form.

Sample: tests_private/comparefiles/Bug693480.pdf

P.S. This was a pain in the butt and I wonder if there are more,
similar, bugs lurking :(